2411101323virus777

+++ Cpp/boj_solve/boj14118_3D_hpi.cpp
@@ -9,8 +9,8 @@
 #include <deque>
 #include <tuple>
 typedef long long ll;
-typedef long double ld;
-//typedef double ld;
+//typedef long double ld;
+typedef double ld;
 typedef std::pair<int, int> pi;
 typedef std::vector<int> Vint;
 typedef std::vector<ld> Vld;
